+++ api-catalog-services/src/main/java/org/zowe/apiml/apicatalog/controllers/api/ApiDocController.java
@@ -31,8 +31,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.zowe.apiml.apicatalog.exceptions.ApiDiffNotAvailableException;
 import org.zowe.apiml.apicatalog.exceptions.ApiDocNotFoundException;
-import org.zowe.apiml.apicatalog.swagger.ApiDocRetrievalService;
-import reactor.core.publisher.Mono;
+import org.zowe.apiml.apicatalog.swagger.ApiDocService;
 
 import static org.apache.hc.core5.http.HttpStatus.SC_OK;
 
@@ -41,12 +40,12 @@ import static org.apache.hc.core5.http.HttpStatus.SC_OK;
  */
 @Slf4j
 @RestController
-@RequestMapping({"/apicatalog/apidoc", "/apicatalog/api/v1/apidoc"})
+@RequestMapping("/apidoc")
 @Tag(name = "API Documentation")
 @RequiredArgsConstructor
 public class ApiDocController {
 
-    private final ApiDocRetrievalService apiDocRetrievalService;
+    private final ApiDocService apiDocService;
 
     /**
      * Retrieve the api-doc info for this service
@@ -69,16 +68,15 @@ public class ApiDocController {
         @ApiResponse(responseCode = "404", description = "URI not found"),
         @ApiResponse(responseCode = "500", description = "An unexpected condition occurred"),
     })
-    public Mono<ResponseEntity<String>> getApiDocInfo(
+    public ResponseEntity<String> getApiDocInfo(
         @Parameter(name = "serviceId", description = "The unique identifier of the registered service", required = true, example = "apicatalog")
         @PathVariable(value = "serviceId") String serviceId,
         @Parameter(name = "apiId", description = "The API ID and version, separated by a space, of the API documentation", required = true, example = "zowe.apiml.apicatalog v1.0.0")
         @PathVariable(value = "apiId") String apiId) {
-        return Mono.fromSupplier(() -> ResponseEntity
+        return ResponseEntity
             .status(SC_OK)
             .contentType(MediaType.APPLICATION_JSON)
-            .body(apiDocRetrievalService.retrieveApiDoc(serviceId, apiId))
-        );
+            .body(apiDocService.retrieveApiDoc(serviceId, apiId));
     }
 
     /**
@@ -100,14 +98,13 @@ public class ApiDocController {
         @ApiResponse(responseCode = "404", description = "URI not found"),
         @ApiResponse(responseCode = "500", description = "An unexpected condition occurred"),
     })
-    public Mono<ResponseEntity<String>> getDefaultApiDocInfo(
+    public ResponseEntity<String> getDefaultApiDocInfo(
         @Parameter(name = "serviceId", description = "The unique identifier of the registered service", required = true, example = "apicatalog")
         @PathVariable(value = "serviceId") String serviceId) {
-        return Mono.fromSupplier(() -> ResponseEntity
+        return ResponseEntity
             .status(SC_OK)
             .contentType(MediaType.APPLICATION_JSON)
-            .body(apiDocRetrievalService.retrieveDefaultApiDoc(serviceId))
-        );
+            .body(apiDocService.retrieveDefaultApiDoc(serviceId));
     }
 
     @GetMapping(value = "/{serviceId}/{apiId1}/{apiId2}", produces = MediaType.TEXT_HTML_VALUE)
@@ -123,7 +120,7 @@ public class ApiDocController {
         @ApiResponse(responseCode = "404", description = "URI not found"),
         @ApiResponse(responseCode = "500", description = "An unexpected condition occurred")
     })
-    public Mono<ResponseEntity<String>> getApiDiff(
+    public ResponseEntity<String> getApiDiff(
         @Parameter(name = "serviceId", description = "The unique identifier of the registered service", required = true, example = "apicatalog")
         @PathVariable(value = "serviceId") String serviceId,
         @Parameter(name = "apiId1", description = "The API ID and version, separated by a space, of the API documentation", required = true, example = "zowe.apiml.apicatalog v1.0.0")
@@ -132,18 +129,17 @@ public class ApiDocController {
         @PathVariable(value = "apiId2") String apiId2) {
 
         try {
-            String doc1 = apiDocRetrievalService.retrieveApiDoc(serviceId, apiId1);
-            String doc2 = apiDocRetrievalService.retrieveApiDoc(serviceId, apiId2);
+            String doc1 = apiDocService.retrieveApiDoc(serviceId, apiId1);
+            String doc2 = apiDocService.retrieveApiDoc(serviceId, apiId2);
             ChangedOpenApi diff = OpenApiCompare.fromContents(doc1, doc2);
             HtmlRender render = new HtmlRender();
             String result = render.render(diff);
             //Remove external stylesheet
             result = result.replace("<link rel=\"stylesheet\" href=\"http://deepoove.com/swagger-diff/stylesheets/demo.css\">", "");
-            return Mono.just(ResponseEntity
+            return ResponseEntity
                 .status(HttpStatus.SC_OK)
                 .header(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
-                .body(result)
-            );
+                .body(result);
         } catch (ApiDocNotFoundException adnfe) {
             throw adnfe;
         } catch (Exception e) {
